requirements/common.txt
asgi-redis==1.4.3
vine==1.3.0
commonmark==0.9.1
django==2.2.20
django-allauth==0.43.0
django-filter==2.4.0
django-import-export==2.4.0
djangorestframework==3.10.3
djangorestframework-expiring-authtoken==0.1.3
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.5.0
django-rest-auth[with_social]==0.9.5
docker-compose==1.27.4
drfdocs==0.0.11
drf-yasg==1.17.0
importlib-metadata<5.0
kubernetes==12.0.1
pika==1.1.0
pickleshare==0.7.5
Pillow==7.1.0
psycopg2-binary==2.8.4
pycurl==7.43.0.6
PyJWT==2.1.0
PyYaml==5.1
proc==1.0
rstr==2.2.6
sendgrid==6.4.8
datadog==0.39.0
confluent-kafka==1.9.2
